requirements.txt
flask>=1.1
jsonify>=0.5
requests>=2.27
google-api-python-client>=1.12
google-auth>=1.25
google-auth-oauthlib>=0.4
google-auth-httplib2>=0.0.4
pandas>=1.2
